+++ src/http/HttpRequestImpl.cpp
@@ -250,7 +250,7 @@ int HttpRequest::Impl::sendData(const void* data, size_t len)
         if (ret > 0) {
             raw_bytes_sent_ += ret;
         }
-        if (finish && ret >= send_len) {
+        if (finish && ret >= static_cast<int>(send_len)) {
             req_complete_ = true;
         }
         return ret;
@@ -338,7 +338,7 @@ int HttpRequest::Impl::sendData(const KMBuffer &buf)
         if (ret > 0) {
             raw_bytes_sent_ += ret;
         }
-        if (finish && ret >= send_len) {
+        if (finish && ret >= static_cast<int>(send_len)) {
             req_complete_ = true;
         }
         return ret;
